edit some folder structure
