+++ client/src/components/account/account.js
@@ -1,9 +1,9 @@
-import React from 'react'
+import React from "react";
 
 export default function account() {
   return (
     <div>
       <p>My Account</p>
     </div>
-  )
+  );
 }
